Extract class Task and its subclasses and class Ui
