+++ src/main/java/BongException.java
@@ -1,0 +1,5 @@
+public class BongException extends Exception {
+    public BongException(String msg) {
+        super(msg);
+    }
+}
