+++ src/main/java/data/bong.txt
@@ -1,3 +1,3 @@
 T | 0 | borrow book
 D | 0 | return book | 2025-10-16 1200
-E | 0 | project meeting | 2025-09-30 1200 to 2025-09-30 1500
+E | 0 | project meeting | 2025-09-30 1200 | 2025-09-30 1500
